Move middlewares in their own module
