+++ code/crates/test/tests/it/wal.rs
@@ -7,11 +7,11 @@ use informalsystems_malachitebft_test::{self as malachitebft_test};
 
 use malachitebft_config::{ValuePayload, VoteSyncMode};
 use malachitebft_core_consensus::LocallyProposedValue;
-use malachitebft_core_types::{NilOrVal, Round, SignedVote};
+use malachitebft_core_types::SignedVote;
 use malachitebft_engine::util::events::Event;
-use malachitebft_test::middleware::Middleware;
-use malachitebft_test::{Address, Height, TestContext, ValueId, Vote};
+use malachitebft_test::TestContext;
 
+use crate::middlewares::{ByzantineProposer, PrevoteNil};
 use crate::{HandlerResult, TestBuilder, TestParams};
 
 #[tokio::test]
@@ -464,7 +464,7 @@ async fn wal_multi_rounds(params: TestParams) {
     let mut test = TestBuilder::<()>::new();
 
     test.add_node()
-        .with_middleware(PrevoteNil)
+        .with_middleware(PrevoteNil::when(|_, round, _| round.as_i64() <= 3))
         .start()
         .wait_until(CRASH_HEIGHT)
         .wait_until_round(3)
@@ -493,65 +493,4 @@ async fn wal_multi_rounds(params: TestParams) {
             },
         )
         .await
-}
-
-#[derive(Copy, Clone, Debug)]
-struct ByzantineProposer;
-
-impl Middleware for ByzantineProposer {
-    fn on_propose_value(
-        &self,
-        _ctx: &TestContext,
-        proposal: &mut LocallyProposedValue<TestContext>,
-        reproposal: bool,
-    ) {
-        use informalsystems_malachitebft_test::Value;
-        use rand::Rng;
-
-        if !reproposal {
-            tracing::warn!(
-                "ByzantineProposer: First time proposing value {:}",
-                proposal.value.id()
-            );
-
-            // Do not change the value if it is the first time we propose it
-            return;
-        }
-
-        // Make up a new value that is different from the one we are supposed to propose
-        let new_value = loop {
-            let new_value = Value::new(rand::thread_rng().gen_range(100..=100000));
-            if new_value != proposal.value {
-                break new_value;
-            }
-        };
-
-        tracing::warn!(
-            "ByzantineProposer: Not re-using previously built value {:} but a new one {:}",
-            proposal.value.id(),
-            new_value.id()
-        );
-
-        proposal.value = new_value;
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-struct PrevoteNil;
-
-impl Middleware for PrevoteNil {
-    fn new_prevote(
-        &self,
-        _ctx: &TestContext,
-        height: Height,
-        round: Round,
-        value_id: NilOrVal<ValueId>,
-        address: Address,
-    ) -> Vote {
-        if round.as_i64() <= 3 {
-            Vote::new_prevote(height, round, NilOrVal::Nil, address)
-        } else {
-            Vote::new_prevote(height, round, value_id, address)
-        }
-    }
 }
